xrootdfs: implement xrootdfs_flush
Motivation:
This is mostly an attempt to catch at least some errors when the file is closed, as xrootdfs_release by definition cannot return an error.
If the file was not opened correctly or for any other reason xrootd cannot actually write the file, the error is invisible a small file, but writes don't actually happen.
The error only surfaces when the cache is flushed, and that does not happen for small files until xrootdfs_release.
Change:
xrootdfs_flush is implemented which calls wcache_flush and checks for errors
Result:
Before xrootdfs_release FUSE calls xrootdfs_flush and an error is reported to the user if wcache_flush does not work.