xrdposix: interpret kXR_Unsupported as ENOTSUP instead of ENOSYS
Motivation
To quote errno.h comment on ENOSYS /*Invalid system call number"*/
> This error code is special: arch syscall entry code will return
> -ENOSYS if users try to call a syscall that doesn't exist.  To keep
> failures of syscalls that really do exist distinguishable from
> failures due to attempts to use a nonexistent syscall, syscall
> implementations should refrain from returning -ENOSYS.

There are many reasons for the server to respond kXR_Unsupported.
Some systems (like FUSE) treat ENOSYS very differently from other errors.
For example, when OPEN returns ENOSYS, it's considered a success (for cases where a filesystem does not need to implement an open operation to work).
When xrootdfs OPEN fails because the server returned kXR_Unsupported, it is an error and should be treated as such.
When FUSE requests an extended attribute and the server responds kXR_Unsupported, meaning that a particular query is not supported,
FUSE interprets it as a missing getxattr syscall and never tries to call getxattr again, even for the supported attributes.
In conclusion FUSE shows that ENOSYS has a specific meaning and can't work as an interpretation of kXR_Unsupported.
Why not use ENOTSUP instead?

Change: map kXR_Unsupported to ENOTSUP when interpreting protocol responses into error codes.

Result:
Will make dealing with kXR_Usupported errors easier and more predictable.
Should not break anything in xrootd code:
Errno is only compared to ENOSYS is oss, not on the network side where it can expect kXR_Unsupported to be ENOSYS.
Errno is never compared to EOPNOTSUPP.
Errno is only compared to ENOTSUP in:
 * XrdFrmAdmin::Chksum, comparison for logging
 * XrdFrmAdmin::ChksumList interrupt printing checksums if error is not ESTALE or ENOTSUP. Probably better this way.
 * XrdFrmAdmin::ChksumPrint, comparison for logging
 * XrdFrmAdmin::ConvTest, logging
 * XrdFrmConfig::Configure
   * looks like filesystem-side, not protocol?
   * if it *is* network-side, makes sense to interpret kXR_Unsupported as ENOTSUP
 * XrdOfsFile::open
   // Create the file. If ENOTSUP is returned, promote the creation to
   // the subsequent open. This is to accomodate proxy support.
   if it *is* networks-side, makes sense to interpret kXR_Unsupported as ENOTSUP
 * XrdSysXAttr::Copy
   * called from XrdOssCopy::Copy
   * if it *is* network-side, makes sense to interpret kXR_Unsupported as ENOTSUP